requirements.txt
pillow
numpy
json
